m01d05: Fibonacci - complete
worked on with Justin F & Justin B but lost the code.  So I started over and came up with this.  learned Decimal Format from here: https://mkyong.com/java/java-display-double-in-2-decimal-points/ which allowed me to use the Golden Ratio and reduce the result to a whole number therefore nullifying the difference between the golden ratio and fibonacci